+++ django_spire/ai/chat/urls/page_urls.py
@@ -2,6 +2,7 @@ from django.urls import path
 
 from django_spire.ai.chat.views import page_views
 
+
 app_name = 'page'
 
 urlpatterns = [