+++ django_spire/knowledge/entry/urls/page_urls.py
@@ -2,6 +2,7 @@ from django.urls import path
 
 from django_spire.knowledge.entry.views import page_views
 
+
 app_name = 'page'
 
 urlpatterns = [