+++ django_spire/knowledge/entry/urls/template_urls.py
@@ -2,6 +2,7 @@ from django.urls import path
 
 from django_spire.knowledge.entry.views import template_views
 
+
 app_name = 'template'
 
 urlpatterns = [