+++ django_spire/knowledge/seeding/seed.py
@@ -1,1 +1,3 @@
+from __future__ import annotations
+
 from django_spire.knowledge.collection.seeding.seed import *