+++ test_project/apps/ai/chat/views.py
@@ -1,7 +1,8 @@
 from __future__ import annotations
 
+from typing import TYPE_CHECKING
+
 from django.template.response import TemplateResponse
-from typing_extensions import TYPE_CHECKING
 
 if TYPE_CHECKING:
     from django.core.handlers.wsgi import WSGIRequest
@@ -9,5 +10,4 @@ if TYPE_CHECKING:
 
 def chat_home_view(request: WSGIRequest) -> TemplateResponse:
     template = 'ai/chat/page/chat_home_page.html'
-
     return TemplateResponse(request, template, context={})