+++ test_project/apps/ordering/intelligence/bots.py
@@ -1,19 +1,16 @@
-from dandy.llm import BaseLlmBot
+from dandy import Bot
 
-from test_project.apps.ordering.intelligence import prompts
-from test_project.apps.ordering.intelligence import intel
+from test_project.apps.ordering.intelligence import intel, prompts
 
 
-class OrderingBot(BaseLlmBot):
-    instructions_prompt = prompts.ordering_instruction_prompt()
-    intel_class = intel.OrderingIntel
+class OrderingBot(Bot):
+    llm_role = prompts.ordering_instruction_prompt()
+    llm_intel_class = intel.OrderingIntel
 
-    @classmethod
     def process(
-            cls,
-            user_input: str
+        self,
+        user_input: str
     ) -> intel.OrderingIntel:
-
-        return cls.process_prompt_to_intel(
+        return self.llm.prompt_to_intel(
             prompt=prompts.ordering_user_input_prompt(user_input)
         )